# backend/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: AI QA Platform API
  version: 0.1.0
servers:
  - url: http://localhost:3000/api
paths:
...
  /projects:
    get:
      summary: List projects
      responses:
        "200":
          description: Project list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      summary: Create project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectRequest"
      responses:
        "201":
          description: Created project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}:
    get:
      summary: Get project by id
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200":
          description: Project detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Project not found
    patch:
      summary: Update project
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectRequest"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Project not found
    delete:
      summary: Delete project
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200":
          description: Deleted project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"
        "404":
          description: Project not found
...
components:
  parameters:
...
    ProjectId:
      name: id
      in: path
      required: true
      schema:
        type: string
...
  schemas:
    Project:
      type: object
      required: [id, name, status, defaultLanguage, createdAt, updatedAt]
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: [string, "null"]
        status:
          type: string
          enum: [ACTIVE, ARCHIVED]
        defaultLanguage:
          type: string
          enum: [vi, en]
        configOverrides:
          type: [object, "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateProjectRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          maxLength: 120
        description:
          type: string
          maxLength: 1000
        defaultLanguage:
          type: string
          enum: [vi, en]
          default: vi
        configOverrides:
          type: object
    UpdateProjectRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 120
        description:
          type: string
          maxLength: 1000
        defaultLanguage:
          type: string
          enum: [vi, en]
        configOverrides:
          type: object
    DeleteResponse:
      type: object
      required: [deleted, id]
      properties:
        deleted:
          type: boolean
        id:
          type: string
...